Added logic and views for first login/password reset.
	app.js
Added get for /resetPassword.
Added logic to check if first login for user.
Added post for changing password on first login, and will be used
  later for resetting password.

	DatabaseHandler
Added first login bool assignment for user for first login.
Created saltQuery() and resetPassword() to allow for resetting a
  users password.

	login.css
Added style for password reset page.

	User.js
Added variables for firstLogin, and resetpass.

	resetPassword.ejs
Created a view to allow for reset passwords.