Finished Logic for first login and 3 month old pw.
	population.sql
Added values to insert for pass_age.

	tables.sql
Added columns for pass_age and security_code.
Columns are used to determin if pass is 3 months old, and to add a
  security code to enter if forgot password.

	app.js
Added checks to see if user needs password reset to all gets.
  This avoids someone trying to skip resetting their password.
Added changing session user class reset pass to false after changing

	DatabaseHandler.js

    attemptLogin()
Added logic to check if password is older than 3 months.

    resetPassword()
Added insertion for password date.

	User.js
Removed firstLogin variable as resetPass works for both that and a
  password older than 3 months.