Improved error logging, fixed bug when querying vendors, and UI changes.
	app.js
Improved error logging by saying what was occuring when error happened.
Increased size of comment showing where app.posts begin to better distinguish where they start.

	DatabaseHandler.js
Added distinct to query of vendors to remove duplicate data in query.
Added order by for all vendor queries so it would be in alphabetical order by vendor name.
Changed price query to be + .99 instead of a $1.  But I will be
  changing this again in the future. Because I just thought of a
  better approach.  nums += nums - math.ceil(nums)

	css files.
UI improvements to position elements better, and not change so much.

	itemSearch.js
Changed validiation as it was working incorrectly.
Removed debugging lines.
Added changes in showing and not showing search wrapper for item name.
Added hiding price search when it was not needed.
Added showing N/A in table when no results.

	vendorSearch.js
Added showing N/A in table when no results.
Changed url to change to appropriate url after app changes on
  previous commit to add parameters when clicking on vendor link in table.

	item.ejs
Added css to html render to hide searching on default option.